# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * EduPlatform API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

export interface Course {
  id: number;
  title: string;
  description: string;
  category: string;
  difficulty: string;
  /** @nullable */
  thumbnailUrl?: string | null;
  totalLessons: number;
  estimatedMinutes: number;
  createdAt: string;
}

export interface CourseInput {
  title: string;
  description: string;
  category: string;
  difficulty: string;
  thumbnailUrl?: string;
  estimatedMinutes: number;
}

export interface Lesson {
  id: number;
  courseId: number;
  title: string;
  description: string;
  /** @nullable */
  videoUrl?: string | null;
  order: number;
  durationMinutes: number;
}

export interface QuizQuestion {
  id: number;
  lessonId: number;
  question: string;
  options: string[];
  order: number;
}

export interface LessonWithQuiz {
  id: number;
  courseId: number;
  title: string;
  description: string;
  /** @nullable */
  videoUrl?: string | null;
  order: number;
  durationMinutes: number;
  quizQuestions: QuizQuestion[];
}

export interface CourseWithLessons {
  id: number;
  title: string;
  description: string;
  category: string;
  difficulty: string;
  /** @nullable */
  thumbnailUrl?: string | null;
  totalLessons: number;
  estimatedMinutes: number;
  createdAt: string;
  lessons: Lesson[];
}

export type CourseStatsCategoryCounts = {[key: string]: number};

export type CourseStatsDifficultyCounts = {[key: string]: number};

export interface CourseStats {
  totalCourses: number;
  totalStudents: number;
  categoryCounts: CourseStatsCategoryCounts;
  difficultyCounts: CourseStatsDifficultyCounts;
}

export interface QuizAnswer {
  questionId: number;
  selectedOption: number;
}

export interface QuizSubmission {
  lessonId: number;
  answers: QuizAnswer[];
}

export interface QuizResult {
  score: number;
  total: number;
  passed: boolean;
  xpEarned: number;
  correctAnswers: number[];
}

export interface UserProgress {
  id: number;
  userId: string;
  courseId: number;
  lessonId: number;
  completed: boolean;
  /** @nullable */
  score?: number | null;
  xp: number;
  /** @nullable */
  completedAt?: string | null;
}

export interface ProgressInput {
  courseId: number;
  lessonId: number;
  completed: boolean;
  score?: number;
  xp: number;
}

export interface ActivityItem {
  lessonTitle: string;
  courseTitle: string;
  completedAt: string;
  xpEarned: number;
}

export interface ProgressDashboard {
  totalXp: number;
  totalCompleted: number;
  currentStreak: number;
  coursesInProgress: number;
  recentActivity: ActivityItem[];
}

export interface LeaderboardEntry {
  rank: number;
  userId: string;
  displayName: string;
  totalXp: number;
  lessonsCompleted: number;
  isCurrentUser: boolean;
}

export interface Leaderboard {
  entries: LeaderboardEntry[];
  currentUserEntry: LeaderboardEntry | null;
}

export interface ChatConversation {
  id: number;
  userId: string;
  title: string;
  createdAt: string;
}

export interface ChatConversationInput {
  title: string;
}

export interface ChatMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface ChatConversationWithMessages {
  id: number;
  userId: string;
  title: string;
  createdAt: string;
  messages: ChatMessage[];
}

export interface ChatMessageInput {
  content: string;
}

export type ListCoursesParams = {
category?: string;
difficulty?: string;
};

export type GetLeaderboardParams = {
limit?: number;
};