# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: EduPlatform API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /progress:
    get:
      operationId: getMyProgress
      tags: [progress]
      summary: Get the current user's progress across all courses
      responses:
        "200":
          description: User progress list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserProgress"
    post:
      operationId: upsertProgress
      tags: [progress]
      summary: Mark a lesson as complete or update progress
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProgressInput"
      responses:
        "200":
          description: Updated progress
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProgress"
...
  /progress/leaderboard:
    get:
      operationId: getLeaderboard
      tags: [progress]
      summary: Get top users by XP with the current user's rank
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Leaderboard with top users and current user rank
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Leaderboard"
...
components:
  schemas:
...
    UserProgress:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        courseId:
          type: integer
        lessonId:
          type: integer
        completed:
          type: boolean
        score:
          type: ["integer", "null"]
        xp:
          type: integer
        completedAt:
          type: ["string", "null"]
          format: date-time
      required:
        - id
        - userId
        - courseId
        - lessonId
        - completed
        - xp
    ProgressInput:
      type: object
      properties:
        courseId:
          type: integer
        lessonId:
          type: integer
        completed:
          type: boolean
        score:
          type: integer
        xp:
          type: integer
      required:
        - courseId
        - lessonId
        - completed
        - xp
...
    LeaderboardEntry:
      type: object
      properties:
        rank:
          type: integer
        userId:
          type: string
        displayName:
          type: string
        totalXp:
          type: integer
        lessonsCompleted:
          type: integer
        isCurrentUser:
          type: boolean
      required:
        - rank
        - userId
        - displayName
        - totalXp
        - lessonsCompleted
        - isCurrentUser
    Leaderboard:
      type: object
      properties:
        entries:
          type: array
          items:
            $ref: "#/components/schemas/LeaderboardEntry"
        currentUserEntry:
          oneOf:
            - $ref: "#/components/schemas/LeaderboardEntry"
            - type: "null"
      required:
        - entries
        - currentUserEntry
...