# lib/api-zod/src/generated/types/leaderboard.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * EduPlatform API specification
 * OpenAPI spec version: 0.1.0
 */
import type { LeaderboardEntry } from './leaderboardEntry';

export interface Leaderboard {
  entries: LeaderboardEntry[];
  currentUserEntry: LeaderboardEntry | null;
}